feat: Add products to Supplier and validations for Supplier lifecycle

+++ src/DevIO.Business/Services/SupplierService.cs
@@ -1,3 +1,4 @@
+using DevIO.Business.Constants;
 using DevIO.Business.Entities;
 using DevIO.Business.Entities.Validations;
 using DevIO.Business.Interfaces.Repositories;
@@ -15,6 +16,12 @@ public class SupplierService(ISupplierRepository supplierRepository) : BaseServi
             return;
         }
 
+        if (await SupplierAlreadyExistsAsync(supplier.Document))
+        {
+            Notify(SupplierValidationMessages.SupplierAlreadyExists);
+            return;
+        }
+
         await supplierRepository.AddAsync(supplier);
     }
 
@@ -25,11 +32,36 @@ public class SupplierService(ISupplierRepository supplierRepository) : BaseServi
             return;
         }
 
+        if (await SupplierAlreadyExistsAsync(supplier.Document, supplier.Id))
+        {
+            Notify(SupplierValidationMessages.SupplierAlreadyExists);
+            return;
+        }
+
         await supplierRepository.UpdateAsync(supplier);
     }
 
     public async Task RemoveAsync(Guid id)
     {
+        var supplier = await supplierRepository.GetSupplierWithProductsAndAddressAsync(id);
+
+        if (supplier is null)
+        {
+            Notify(SupplierValidationMessages.SupplierDotNotExists);
+            return;
+        }
+
+        if (supplier.Products!.Any())
+        {
+            Notify(SupplierValidationMessages.SupplierHasProducts);
+            return;
+        }
+
+        if (supplier.Address is not null)
+        {
+            await supplierRepository.RemoveAddressAndSupplierAsync(supplier.Address);
+        }
+
         await supplierRepository.RemoveAsync(id);
     }
 
@@ -46,6 +78,20 @@ public class SupplierService(ISupplierRepository supplierRepository) : BaseServi
 
     private bool IsSupplierAndAddressValid(Supplier supplier) =>
         IsSupplierValid(supplier) && IsAddressValid(supplier.Address);
+
+    private async Task<bool> SupplierAlreadyExistsAsync(string? document)
+    {
+        var suppliers = await supplierRepository.FindAsync(s => s.Document == document);
+
+        return suppliers.Any();
+    }
+
+    private async Task<bool> SupplierAlreadyExistsAsync(string? document, Guid id)
+    {
+        var suppliers = await supplierRepository.FindAsync(s => s.Document == document && s.Id != id);
+
+        return suppliers.Any();
+    }
 
     #endregion
 }
